requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
botocore==1.31.23
click==8.1.7
colorama==0.4.4
docutils==0.16
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.0.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==0.13.1
psycopg2==2.9.8
psycopg2-binary==2.9.7
py==1.11.0
pyasn1==0.5.0
PyJWT==1.7.1
pytest==6.2.2
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2023.3
PyYAML==6.0.1
rsa==4.7.2
ruamel.yaml==0.16.5
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.20
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
WTForms==3.0.1
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1